+++ src/interfaces/INetworkClient.ts
@@ -39,4 +39,14 @@ export interface INetworkClient {
      * @returns Promise which resolves to the object returned or rejects with error.
      */
     postJson<T, U>(data: T, additionalPath?: string, additionalHeaders?: { [header: string]: string }): Promise<U>;
+
+    /**
+     * Perform a request asynchronously.
+     * @param method The method to send the data with.
+     * @param data The data to send.
+     * @param additionalPath An additional path append to the endpoint path.
+     * @param additionalHeaders Extra headers to send with the request.
+     * @returns Promise which resolves to the object returned or rejects with error.
+     */
+    doRequest(method: string, data: string, additionalPath?: string, additionalHeaders?: { [header: string]: string }): Promise<string>;
 }