+++ src/loggers/consoleLogger.ts
@@ -14,6 +14,8 @@ export class ConsoleLogger implements ILogger {
 
     /**
      * Create and instance of the console logger.
+     * @param loggingObject The object to send all the logging to.
+     * @returns A new instance of ConsoleLogger.
      */
     constructor(loggingObject?: Console) {
         this._loggingObject = loggingObject || console;