+++ src/helpers/stringHelper.ts
@@ -37,7 +37,7 @@ export class StringHelper {
      * @returns The escaped version of the string.
      */
     public static encodeNonASCII(value: string): string {
-        return StringHelper.isString(value) ? value.replace(/[\u007F-\uFFFF]/g, (chr) => `\\u${(`0000${chr.charCodeAt(0).toString(16)}`).substr(-4)}`) : undefined;
+        return StringHelper.isString(value) ? value.replace(/[\u0100-\uFFFF]/g, (chr) => `\\u${(`0000${chr.charCodeAt(0).toString(16)}`).substr(-4)}`) : undefined;
     }
 
     /**